Simplify how we communicate with the engineproxy binary
** BACKWARDS-INCOMPATIBLE CHANGES **

- The `logger` option to `new Engine` added recently in #114 no longer
  exists. It is replaced by `proxyStdoutStream` and `proxyStderrStream` options,
  as well as a `restarting` event on the `Engine` object.

- The default log style is now the same as in the Docker container release of
  engineproxy: textual logs over stdout, instead of JSON over stderr.

** END BACKWARDS-INCOMPATIBLE CHANGE NOTES **

Previously, we relied on parsing engine logs to wait until it is listening and
discover what address it is listening on.

This led to a lot of complexity. We hardcoded engineproxy's log configuration to
to be JSON/stdout (which isn't even the engineproxy default of text/stderr) and
did an elaborate dance to ensure that INFO messages temporarily get into
apollo-engine even if the user only wanted higher log level messages. This means
we always modified the incoming config (even for single proxy mode) and needed
to use the more complex -config=stdin so that we could re-configure the proxy.

Instead, a new version of engineproxy prints its listen address to a non-stdio
fd if the $LISTENING_REPORTER_FD environment variable is set. Now we don't have
to mess with logging configuration at all, and in fact by default we can have
engineproxy write directly to our stdout/stderr instead of having this process
waste time copying bytes around (and perhaps failing to do so if it crashes). We
can set the config via $ENGINE_CONFIG and not have to interact with a stdin pipe
at all.

Because the default is to no longer send stdout through a console.log-style API,
this change removes the custom JS logger feature added recently in #114. Instead
of passing in a console-style `logger` object with log and error methods, you can
pass in writable streams `proxyStdoutStream` and `proxyStderrStream`. One
particular message that was accessible via the custom logger's `error` method is
now emitted as a `restarting` event on the Engine object.