devDependencies: update apollo-server-*
Notably, this gets us a version of apollo-server-micro whose peerDependencies on
micro matches the version of micro we have here, so we don't get a warning on
`npm install`.

However, we keep apollo-server-micro at 1.2.x, because 1.3.x only works with the
newer hapi v17, and apollo-engine-js does not yet support hapi v17 (#85).