New JS API (#150)
See https://meteor.quip.com/einIA4ozsUHX/Proposal-for-a-new-apollo-engine-JS-API
for motivations.

https://www.apollographql.com/docs/engine/ is being revamped to match.

This removes "double proxying" and framework-specific middleware, and replaces
it with a wrapper around 'listen'. It depends on changes to engineproxy. It
renames the main entry point from Engine to ApolloEngine.

It adds support for Restify and Hapi 17.

It also adds an ApolloEngineLauncher API which has no Node web framework
integration, as an alternative to Docker for folks whose GraphQL servers aren't
written in Node but still can run Node.

Also rewrite tests from Mocha/JS to Jest/TS.